# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Получить список всех авторов с книгами",
                "operationId": "c66389cfbe7eecf32fad2f1b1af8da7b",
                "responses": {
                    "200": {
                        "description": "Список авторов с книгами",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Добавить нового автора",
                "operationId": "1ccb37eb2efc8b9e14d7632d1de576ac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Имя автора",
                                        "type": "string",
                                        "example": "Иван Тургенев"
                                    },
                                    "bio": {
                                        "description": "Биография автора",
                                        "type": "string",
                                        "example": "Русский писатель и драматург."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Автор успешно добавлен",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Неверные данные запроса"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/v1/authors/{id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Получить информацию об авторе",
                "operationId": "fa6937c010a766599466bc0fbe65dad4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID автора",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об авторе",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Автор не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Authors"
                ],
                "summary": "Обновить информацию о авторе",
                "operationId": "5caa5b324f5c5ec2716297cb69e1d868",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID автора",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Имя автора",
                                        "type": "string",
                                        "example": "Иван Тургенев"
                                    },
                                    "bio": {
                                        "description": "Биография автора",
                                        "type": "string",
                                        "example": "Русский писатель и драматург."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Автор успешно обновлен",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Неверные данные запроса"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Удалить автора",
                "operationId": "80da65b3b24823d566c9fa5d26f545bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID автора",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Автор удален"
                    },
                    "404": {
                        "description": "Автор не найден"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}